Renamed `syntax` module to `ast`.
